apps/worker/pyproject.toml
[project]
name = "worker-app"
version = "0.1.0"
description = "Background worker for EduAssist Platform."
requires-python = ">=3.12"
dependencies = [
  "llama-index-core>=0.13.0,<0.14.0",
  "llama-index-vector-stores-qdrant>=0.7.0,<0.8.0",
  "minio>=7.2.8",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.8.0",
  "PyYAML>=6.0.2",
  "qdrant-client[fastembed]>=1.14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/worker_app"]
